refactor: ProfileForm
- Relocated FormInput module
- Cleaned up unused imports from FormInput.jsx
- Removed GenerateReferralLink from routes and Navbar
- Added new icons: Location, Phone, Edit
- Introduced Loader component
- Created ProfileForm for user profile management
- Implemented ProfilePage with new ProfileForm and refactored layout
- Enhanced ExercisesPage with Loader and modal for creating exercises
